Move the session under services folder.
- E2E test can mock the whole services to mock the session function.
